requirements.txt
attrs==25.3.0
Automat==24.8.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
constantly==23.10.4
cryptography==44.0.2
cssselect==1.3.0
defusedxml==0.7.1
filelock==3.18.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
itemadapter==0.11.0
itemloaders==1.3.2
jmespath==1.0.1
lxml==5.3.1
numpy==2.0.2
packaging==24.2
parsel==1.10.0
pillow==11.1.0
Protego==0.4.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
PyDispatcher==2.0.7
pyOpenSSL==25.0.0
queuelib==1.8.0
requests==2.32.3
requests-file==2.1.0
scipy==1.13.1
Scrapy==2.12.0
service-identity==24.2.0
tldextract==5.1.3
tomli==2.2.1
Twisted==24.11.0
typing_extensions==4.13.1
urllib3==2.3.0
w3lib==2.3.1
zope.interface==7.2
